refactor: getallPost request from local to global
