+++ src/context-api/PostProvider.js
@@ -1,4 +1,6 @@
 import { useContext, createContext, useReducer } from "react";
+import React, { useEffect, useCallback } from "react";
+import axios from "axios";
 
 const PostContext = createContext();
 const reducer = (state, action) => {
@@ -46,6 +48,22 @@ const initialState = {
 };
 export const PostProvider = ({ children }) => {
   const [state, postDispatch] = useReducer(reducer, initialState);
+  const getAllPosts = useCallback(async () => {
+    try {
+      const { data, status } = await axios.get(
+        `${process.env.REACT_APP_API_URL}/posts`
+      );
+      if (data.success && status === 200) {
+        const posts = data.posts.reverse()
+        postDispatch({ type: "LOAD_POST_DATA", payload: posts });
+      }
+    } catch (err) {
+      console.log(err.response);
+    }
+  }, [postDispatch]);
+  useEffect(() => {
+    getAllPosts();
+  }, [getAllPosts]);
   return (
     <PostContext.Provider value={{ ...state, postDispatch }}>
       {children}
