+++ src/posts/components/Posts.js
@@ -1,5 +1,4 @@
-import axios from "axios";
-import React, { useEffect, useCallback } from "react";
+import React from "react";
 import { usePostContext } from "../../context-api/PostProvider";
 import PostCard from "./PostCard";
 import { ContentLoader } from "../../shared/components/Loader";
@@ -7,27 +6,11 @@ import { ContentLoader } from "../../shared/components/Loader";
 import "./Posts.css";
 
 function Posts() {
-  const { posts, postDispatch } = usePostContext();
-  const getAllPosts = useCallback(async () => {
-    try {
-      const { data, status } = await axios.get(
-        `${process.env.REACT_APP_API_URL}/posts`
-      );
-      if (data.success && status === 200) {
-        const posts = data.posts.reverse()
-        postDispatch({ type: "LOAD_POST_DATA", payload: posts });
-      }
-    } catch (err) {
-      console.log(err.response);
-    }
-  }, [postDispatch]);
-  useEffect(() => {
-    getAllPosts();
-  }, [getAllPosts]);
+  const { posts } = usePostContext();
+  
   return (
     <>
-      {posts.length === 0 && <ContentLoader />}
-      {posts && posts.map((post) => <PostCard data={post} key={post._id} />)}
+    {posts.length === 0 ? <ContentLoader/> : posts?.map((post) => (<PostCard data={post} key={post._id} />))}
     </>
   );
 }
